pgw#1206 C1: the store outlives the job, so it leaves the job engine
`ModelStore` (1,668 lines, 55 methods) moves out of `executor.py` into
`gen_worker/models/store.py`. Phase C splits the executor BY LIFETIME, and the
store is the longest-lived thing a worker owns — a pod's store outlives every
job that runs on it — so it splits off first.

executor.py 12,873 -> 11,062 lines. This is a move, not a rewrite: no method
body changed. Two forced deltas, both stated rather than smuggled:

* `_sanitize` becomes `gen_worker.redact.sanitize` (25-line module). The store
  may not import the executor, and both sides redact outbound messages, so the
  helper needed a neutral home rather than a duplicate or a re-export.
* three bare `set` annotations become `Set[str]`. Strict mypy demands type
  arguments of a new module; executor.py's ratchet entry had been covering
  them. Fixed properly rather than by growing the ratchet.

What moved with it, because nothing else uses it: `_MaterializedLocal`,
`_is_terminal_download_error`, `_snapshot_to_resolved`,
`_snapshot_without_components`, `_snapshot_files_without_components`, the
residency/tier pb maps, `_ResidencyIdentity`, `_USE_RESIDENT_IDENTITY` and the
five download/disk-GC constants. `_GiB` is redefined locally, matching what
preload.py, runtimes/llama.py and utils/lora.py already do.

Importers move to the real path rather than leaning on the incidental
re-export `executor.py` still has (it constructs a store): 24 files, no shim
and no alias. Tests that patched `executor.ensure_local` / `verify_files` /
`_PROGRESS_EVENT_MIN_INTERVAL_S` now patch `gen_worker.models.store` — the spy
belongs at the DEFINING module (pgw#1210).

The `RUNPOD_POD_ID` / `RUNPOD_PROVIDER` / `TENSORHUB_FILL_SOURCE_DIR` config
reads are repointed in both `config_reads_allowlist.txt` and the
`BEHAVIOUR_GATES` table. Their classification and burn-down blockers are
unchanged: C1 moved the reads, it did not license them.

Fence: this touches none of the 17 methods frozen for th#1834's Phase 3
rewrite (executor.py 9424-10525 at 73f826e1) and none of Phase 3's edit sites.
`_background_mint_run` / `_delegated_mint_run` / `_advertise_minted_cells`
remain adjacent as they asked.

Verified: strict mypy clean (759 files), ruff clean, all 16 fast-gate lint
scripts green, 220 tests across every affected file green.
