+++ changelog.d/pgw1206.md
@@ -1,0 +1,12 @@
+- **BREAKING (pgw#1206 C1) — `ModelStore` leaves `executor.py`.**
+  `gen_worker.executor.ModelStore` → `gen_worker.models.store.ModelStore`. Phase
+  C splits the executor by LIFETIME, and the store is the longest-lived thing a
+  worker owns — a pod's store outlives every job on it — so it goes first;
+  `executor.py` keeps the job engine and drops from 12,873 to 11,062 lines. A
+  verbatim move: no method body changed. The only content deltas are three
+  `set` → `Set[str]` annotations that strict mypy requires of a new module, and
+  `_sanitize` becoming `gen_worker.redact.sanitize` because the store may not
+  import the executor. Tests and harnesses that patched `executor.ensure_local`,
+  `executor.verify_files`, `executor._snapshot_to_resolved` or
+  `executor._PROGRESS_EVENT_MIN_INTERVAL_S` must patch `gen_worker.models.store`
+  instead — the spy belongs at the DEFINING module (pgw#1210).
