+++ tests/test_boot_adopt_credential_pgw1108.py
@@ -39,7 +39,8 @@ class _Spec:
 
 
 def _executor(tmp_path: Path) -> Any:
-    from gen_worker.executor import Executor, ModelStore
+    from gen_worker.executor import Executor
+    from gen_worker.models.store import ModelStore
 
     async def _send(msg: Any) -> None:
         pass
