+++ tests/test_boot_adopt_local_first_pgw1127.py
@@ -178,7 +178,8 @@ class _Arm:
 
 
 def _executor(tmp_path: Path) -> Any:
-    from gen_worker.executor import Executor, ModelStore
+    from gen_worker.executor import Executor
+    from gen_worker.models.store import ModelStore
 
     async def _send(msg: Any) -> None:
         pass
