+++ tests/test_component_digests_v2_th1303.py
@@ -15,7 +15,7 @@ from __future__ import annotations
 
 from pathlib import Path
 
-from gen_worker.executor import ModelStore
+from gen_worker.models.store import ModelStore
 from gen_worker.pb import worker_scheduler_pb2 as pb
