+++ tests/test_hello_ack_model_diff_gw614.py
@@ -20,7 +20,8 @@ from pathlib import Path
 from types import SimpleNamespace
 
 from gen_worker.config import Settings
-from gen_worker.executor import Executor, ModelStore
+from gen_worker.executor import Executor
+from gen_worker.models.store import ModelStore
 from gen_worker.lifecycle import Lifecycle
 from gen_worker.pb import worker_scheduler_pb2 as pb
