+++ tests/test_intent_registry_th1085.py
@@ -7,7 +7,8 @@ from types import SimpleNamespace
 import msgspec
 import pytest
 
-from gen_worker.executor import Executor, ModelStore
+from gen_worker.executor import Executor
+from gen_worker.models.store import ModelStore
 from gen_worker.lifecycle_intents import IntentRegistry, UnreportedIntentWait
 from gen_worker.pb import worker_scheduler_pb2 as pb
 from gen_worker.registry import EndpointSpec
