+++ tests/test_mint_reopen_pgw677.py
@@ -56,9 +56,11 @@ from gen_worker import (
 from gen_worker import compile_cache as cc
 from gen_worker import fleet_cells, hot_swap, mint_delegate
 from gen_worker.api.binding import Hub, wire_ref
-from gen_worker.executor import Executor, ModelStore
+from gen_worker.executor import Executor
+from gen_worker.models.store import ModelStore
 from gen_worker.pb import worker_scheduler_pb2 as pb
 from gen_worker.registry import extract_specs
+from gen_worker.models import store as store_mod
 
 FAMILY = "sdxl"
 _ASPECT_AXIS = CompileAxis(classes=(
@@ -190,7 +192,7 @@ class _Harness:
             p.mkdir(parents=True, exist_ok=True)
             return p
 
-        monkeypatch.setattr(executor_mod, "ensure_local", _fake_ensure_local)
+        monkeypatch.setattr(store_mod, "ensure_local", _fake_ensure_local)
         monkeypatch.setattr(
             fleet_cells, "enable_compiled", self._fake_enable_compiled)
         # pgw#1010: every mint is a CHILD mint now (the in-process capture only
