+++ tests/test_receipt_identity_seam_pgw1122.py
@@ -431,7 +431,8 @@ def _hit(family: str = FAMILY, function: str = "generate") -> Any:
 
 
 def _executor(tmp_path: Path) -> Any:
-    from gen_worker.executor import Executor, ModelStore
+    from gen_worker.executor import Executor
+    from gen_worker.models.store import ModelStore
 
     async def _send(msg: Any) -> None:
         pass
