+++ tests/test_scope_armed_slot_target_pgw1078.py
@@ -34,9 +34,11 @@ import gen_worker.executor as executor_mod
 from gen_worker import Compile, Resources
 from gen_worker import compile_cache as cc
 from gen_worker.api.binding import Hub, wire_ref
-from gen_worker.executor import Executor, ModelStore
+from gen_worker.executor import Executor
+from gen_worker.models.store import ModelStore
 from gen_worker.pb import worker_scheduler_pb2 as pb
 from gen_worker.registry import EndpointSpec
+from gen_worker.models import store as store_mod
 
 FAMILY = "minimax-h3"
 
@@ -72,7 +74,7 @@ def _executor(spec: EndpointSpec, tmp_path: Path, sent: list, monkeypatch) -> Ex
     async def _fake_ensure_local(ref, **kwargs) -> Path:
         return tmp_path / "snap"
 
-    monkeypatch.setattr(executor_mod, "ensure_local", _fake_ensure_local)
+    monkeypatch.setattr(store_mod, "ensure_local", _fake_ensure_local)
     return Executor([spec], _send, store=store)
